narrowDown: filter out flavor text from AI: "Here are the titles", etc.
In narrowDownPageNames(), the only thing we want to hear from AI is the
list of page names. Sometimes AI prepends it with "Here are the titles
most likely to contain relevant content", sometimes it doesn't.
Either way, everything other than page names should be filtered out.

In the current response of ChatGPT, page titles are prepended with "-"
(likely unordered list in Markdown?), so we can check for that.