findpar.js: discard matches if snippet consists of too many parts
If a snipet was generated from list or table, not regular text,
this snippet can have numerous consecutive words that are not
a part of contiguous text.
Attempting to calculate paragraph on such a snippet is pointless
(would produce many irrelevant paragraph numbers),
so we discard the matches if snippet was cut into too many parts.